requirements.txt
tensorflow==2.16.1
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.1
pyspark==3.4.1
PyYAML==6.0.1
scikit_learn==1.4.2
typer==0.12.3
